+++ comet_pqc/preferences/__init__.py
@@ -1,0 +1,3 @@
+from .table import TableTab
+from .webapi import WebAPITab
+from .options import OptionsTab